+++ src/main/java/com/openggf/tools/TraceReplayDrive.java
@@ -246,16 +246,6 @@ final class TraceReplayDrive {
         }
 
         @Override
-        public void advancePlayableAnimationsOnly() {
-            driver.advancePlayableAnimationsOnly();
-        }
-
-        @Override
-        public void advancePlayableFixedSlotsOnly() {
-            driver.advancePlayableFixedSlotsOnly();
-        }
-
-        @Override
         public void suppressFirstSidekickAnimationOnce() {
             driver.suppressFirstSidekickAnimationOnce();
         }