+++ src/test/java/com/openggf/tests/HeadlessTestFixture.java
@@ -81,16 +81,6 @@ public final class HeadlessTestFixture implements TraceReplayFixture {
     }
 
     @Override
-    public void advancePlayableAnimationsOnly() {
-        runner.advancePlayableAnimationsOnly();
-    }
-
-    @Override
-    public void advancePlayableFixedSlotsOnly() {
-        runner.advancePlayableFixedSlotsOnly();
-    }
-
-    @Override
     public void suppressFirstSidekickAnimationOnce() {
         runner.suppressFirstSidekickAnimationOnce();
     }